# store-service/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/point": {
            "get": {
                "description": "Get user's total point balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "point"
                ],
                "summary": "Get total points",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/point.Point"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Deduct points from user's point balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "point"
                ],
                "summary": "Deduct points from user",
                "parameters": [
                    {
                        "description": "Point deduction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/point.SubmitedPoint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/point.Point"
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/point/calculate": {
            "get": {
                "description": "Calculate how many points a payment amount earns (delegated to point-service)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "point"
                ],
                "summary": "Calculate receive points from an amount",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Payment amount in THB",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/point.TotalPoint"
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "point.Point": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "point.SubmitedPoint": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                }
            }
        },
        "point.TotalPoint": {
            "type": "object",
            "properties": {
                "point": {
                    "type": "integer"
                }
            }
        },
...
    }
}